Isolation: fixed mapping of user and group IDs.
The IDs are retrieved in the host user namespace, and so they should be
compared to the "host" part of the mappings.

Also, since they are retrieved in the "host" user namespace, we need
to apply the mappings to them before using them in set[ug]id(2), which
we call in the "container" user namespace.

Closes: <https://github.com/nginx/unit/issues/788>
Reported-by: <bsidhu009@gmail.com>
Cc: Tiago Natel de Moura <t.nateldemoura@f5.com>
Signed-off-by: Alejandro Colomar <alx@nginx.com>