Isolation: fixed unsafe code.
Calling snprintf(3) for concatenating strings is _unsafe_.  The affected
code in this patch has had a bug in two different implementations before
finally being fixed.  That can only be attributed to the lack of an
appropriate API in the C standard library for copying strings with
truncation.  snprintf(3) was being used because there is nothing better.

Let's use the newly added nxt_stpecpy(), which has a simpler interface.

As a side effect, this optimizes the code, since this API is optimal
for copying strings with truncation (for example, it removes calls to
strlen(3); not calling formatting functions also removes overhead).

Closes: <https://github.com/nginx/unit/issues/804>
Signed-off-by: Alejandro Colomar <alx@nginx.com>