Router: refactored code.
This is a preparation for the next commit, which will add logging.
Since we need to add an extra parameter to many functions to be able to
log, and also need to be able to call a logging function just before
returning, and that causes a huge diff, it's better to break the no-op
changes into a separate patch, for improving readability of the actual
changes.

Reviewed-by: Andrew Clayton <a.clayton@nginx.com>
Signed-off-by: Alejandro Colomar <alx@nginx.com>