Tools: using nicer characters for showing a tree.
Especially in small trees, ASCII characters are confusing.  Use nicer
UTF-8 characters, which are more readable to the audience of this
script.  We don't expect the audience of this script to have limited
environments where these characters will not be shown, but if that
happens, we could improve the script to select the caracters based on
the locale.

Suggested-by: Liam Crilly <liam@nginx.com>
Reviewed-by: Andrew Clayton <a.clayton@nginx.com>
Signed-off-by: Alejandro Colomar <alx@nginx.com>