Log: added "diag" log level.
This log level is intended for per-request entries that are non-"debug".

We have a need for logging per-request information to enable some level
of debugging of the routes cofiguration.  This needs to be available in
non-debug mode, since users may want to analyze their running instances
without having to restart unitd with a build of unitd with debugging
enabled.

The info level is enabled by default, so it's not suitable for that,
since logging a few lines per request for many thousands or millions of
requests would be very noisy for the log.  Also, I/O would degrade
performance considerably on default builds of unitd.

But debug is not good for some per-request logs, because they're not
intended to debug unitd, but rather the configuration and how it
interprets the incoming requests.

So this new "diag" level will be the in-between level, intended to be
high-traffic, and only enabled manually.  It will be something available
for all builds of unitd.

We intend to provide a mechanism to enable and disable the log level at
run time.

Suggested-by: Liam Crilly <liam@nginx.com>
Reviewed-by: Andrew Clayton <a.clayton@nginx.com>
Signed-off-by: Alejandro Colomar <alx@nginx.com>