Configuration: added "settings/http/log_route_selection".
This adds configurability to the error log.  For now, it only has one
toggle: "settings/http/log_route_selection".  It allows enabling and
disabling logs related to how the router performs selection of the
routes.

The macros nxt_log_http() and nxt_debug_http() accept a "condition",
which can be an OR of various features, and if any of them are enabled
in the configuration, the log will be printed.  This OR will be useful
for example for printing the request line, which we don't want to print
by default, but we will want to print it if we're printing any of the
conditional logs that can be printed in a request (right now there's
only one, "log_route_selection", but in the future we will likely add at
least one more for the URI rewrite feature).

Link: <https://github.com/nginx/unit/issues/758>
Link: <https://github.com/nginx/unit/pull/824>
Link: <https://github.com/nginx/unit/pull/839>
Suggested-by: Timo Stark <t.stark@nginx.com>
Suggested-by: Liam Crilly <liam@nginx.com>
Cc: Artem Konev <a.konev@f5.com>
Cc: Zhidao Hong <z.hong@f5.com>
Signed-off-by: Alejandro Colomar <alx@nginx.com>

http