Configuration: added "settings/log".
This adds configurability to the error log.  For now, it only has one
child: "settings/log/level".  It allows changing the log level at run
time.

This configurable log level only affects router logging, and doesn't
affect the main, controller, and application processes.  The application
processes can't be configured, since they ignore the log level (except
that they print the log level as a prefix to the log message, but they
don't log conditionally).  The main and controller processes should be
able to be modified in that regard.  However, since it's more complex
than just modifying the router, and they don't log too much, I don't
think it's essential to implement that now.

I implemented it numerically, since that is more intuitive to users as
in "higher numbers print more stuff", which is usual in virtually all
software.  Using the log level textual name might be less clear, since
one needs to remember in which order they go.

Cc: Liam Crilly <liam@nginx.com>
Cc: Artem Konev <a.konev@f5.com>
Cc: Zhidao Hong <z.hong@f5.com>
Cc: Andrew Clayton <a.clayton@nginx.com>
Signed-off-by: Alejandro Colomar <alx@nginx.com>