Change nxt_bool_t from a nxt_uint_t to a uint8_t.
We have a 'bool' type, nxt_bool_t which is defined as a nxt_uint_t, that
is used as local variables, and function return types and arguments.

However in structs we use a uint8_t, due to not wanting to waste memory
in structures, but the use of uint8_t in structs is confusing and
unintuitive.

We could just switch to using the C99 _Bool type, however there are
concerns about this possibly not always being 1-byte on platforms on
which Unit runs and might mess with structure layouts.

(Switching to _Bool as the underlying type is still an eventual goal)

The alternative is to just make our nxt_bool_t be a uint8_t.

Reviewed-by: Alejandro Colomar <alx@nginx.com>
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>