Convert uint8_t struct boolean members to nxt_bool_t.
Replace the usage of uint8_t in structures to represent boolean values
with our nxt_bool_t type.

This will result in no change in structure layout as the nxt_bool_t is
now a uint8_t, same as what it's replacing.

Even though it's essentially the same type, it makes it much clearer as
to what its purpose is.

This was largely done with the following script from Alex, with some
manual conversions

  $ grep -rl 'uint8_t.*1 bit' src/ \
  | xargs sed -i '/uint8_t.*1 bit/{s/uint8_t     /nxt_bool_t  /;s/; *\/\*.*/;/}'

This doesn't convert the non-uint8_t booleans, they will be handled
separately.

Reviewed-by: Alejandro Colomar <alx@nginx.com>
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>