Add NXT_CONF_MAP_BOOL.
When adding a boolean config option, this is of type
NXT_CONF_VALUE_BOOLEAN, however when mapping such an entry to a
structure member via nxt_conf_map_object() this is done as a
NXT_CONF_MAP_INT8. It took a bit of head scratching to find out that it
should have been a NXT_CONF_MAP_INT8 and not a NXT_CONF_MAP_INT.

Introduce a new map type of NXT_CONF_MAP_BOOL that will eventually
replace NXT_CONF_MAP_INT8 for representing booleans...

We use the short form _BOOL as that matches the naming convention of the
other types in nxt_conf_map_type_t.

This is part of a set of patches and eventually NXT_CONF_MAP_BOOL will
simply replace NXT_CONF_MAP_INT8.

Reviewed-by: Alejandro Colomar <alx@nginx.com>
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>