Use nxt_bool_t in the configuration type system.
This replaces the use of uint8_t to represent booleans in the
configuration type system with the nxt_bool_t type (which happens to be
a uint8_t).

This is part of the work to use nxt_bool_t in structures to represent
booleans rather than uint8_t's which is more intuitive.

Reviewed-by: Alejandro Colomar <alx@nginx.com>
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>