Convert uint32_t struct boolean members to nxt_bool_t.
This commit complements the earlier one that changed a bunch of uint8_t
struct members to nxt_bool_t (aka uint8_t).

It's not entirely clear why these were specifically done as uint32_t's.

I've checked these structure layouts with pahole(1) after this change on
both 32 and 64bit and the layouts haven't changed.

We may get a 3 or 7 byte hole after the nxt_bool_t or 3 bytes of padding
at the end of the structure, but the structure members are still at the
same offsets and the structure sizes have remained the same.

Reviewed-by: Alejandro Colomar <alx@nginx.com>
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>