Docs: setting the project name in unitd.8.
Not setting it produces the default value of 'BSD' or 'GNU', depending
on the software formatting the manual page.  We're neither, so let's
specify our project name.  See groff_mdoc(7).  While mandoc_mdoc(7)
formally says that .Os is only for the operating system, and not for the
package name, that's an oversimplification, and only meant for software
inherent to the OS.  For portable software, mandoc(1)'s (and OpenBSD's)
maintainer Ingo Schwarze agreed that it is more sensible to specify the
project name (and optionally, the version).

Reviewed-by: Artem Konev <a.konev@f5.com>
Signed-off-by: Alejandro Colomar <alx@nginx.com>